Add base85.
Special handling of all zeroes and all spaces is not implemented yet.  A good
way of dealing with the current space leaks is crucial for it--what is a few
extra bytes in other encodings turns into swaths of memory for base85
when considering decoding of a string of all y's.

Signed-off-by: Magnus Therning <magnus@therning.org>