Handle special chars in base85.
Adding handling of special chars ('z' = 4 zeroes, 'y' = 4 spaces) in base85.
This also required handling insufficient space in the destination buffer,
which is done similarly to yenc.  Just like there it's sometimes necessary to
iterate over the input in case the output doesn't fit in the provided buffer.
In the case of base85 that only happens for decoding.

Adding proper handling of zero length source as well.

Signed-off-by: Magnus Therning <magnus@therning.org>